refactor(harness): collapse transcript session controllers
